Add handler test.
